+++ client/src/components/system/TopBar.js
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import { useQuery } from "react-query";
 import { Link } from "react-router-dom";
 
-import connectionAnim from "../../images/connection_status.png";
+import connectionAnim from "../../images/sprites/network-status.png";
 import SpriteSheet from "../animations/SpriteSheet";
 import Avatar from "../Avatar";
