+++ client/src/images/sprites/fight-loop.png
