+++ client/src/images/sprites/loading-loop.png
