+++ client/src/images/sprites/network-status.png
