+++ client/src/images/sprites/waiting-action.png
