+++ client/src/images/sprites/waiting-loop.png
